+++ qualang_tools/wirer/connectivity/types.py
@@ -1,7 +1,9 @@
 from typing import Union, List, Tuple
 
-QubitType = int
-QubitPairType = Tuple[int, int]
+ElementType = Union[int, str]
+QubitType = Union[int, str]
+QubitPairType = Union[Tuple[int, int], Tuple[str, str], Tuple[int, str]]
 
+ElementsType = Union[ElementType, List[ElementType]]
 QubitsType = Union[QubitType, List[QubitType]]
 QubitPairsType = Union[QubitPairType, List[QubitPairType]]